+++ src/utils/base.ts
@@ -1,4 +1,5 @@
 import { isEmpty } from "./is";
+import { jsonEncode } from "./parse"
 
 export function timestamp() {
   return new Date().getTime();
@@ -85,24 +86,6 @@ export function toSet<T = any>(datas: T[], byKey?: (e: T) => any) {
     return newDatas;
   }
   return Array.from(new Set(datas));
-}
-
-export function jsonEncode(obj: any, options?: { prettier?: boolean }) {
-  const { prettier } = options ?? {};
-  try {
-    return prettier ? JSON.stringify(obj, undefined, 4) : JSON.stringify(obj);
-  } catch (error) {
-    return undefined;
-  }
-}
-
-export function jsonDecode(json: string | null | undefined) {
-  if (json == undefined) return undefined;
-  try {
-    return JSON.parse(json!);
-  } catch (error) {
-    return undefined;
-  }
 }
 
 export function withDefault<T = any>(e: any, defaultValue: T): T {
